+++ lib/l10n/app_localizations_en.dart
@@ -1435,6 +1435,12 @@ class AppLocalizationsEn extends AppLocalizations {
   String get xenRoot => 'ROOT';
 
   @override
+  String get xenFamily => 'FAMILY';
+
+  @override
+  String get xenScale => 'SCALE';
+
+  @override
   String get xenLatch => 'LATCH';
 
   @override